% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{apprSVMParetoFronts}
\alias{apprSVMParetoFronts}
\title{Performance of different approximative SVM solvers}
\format{A data frame with 994 rows and 5 variables:
\describe{
  \item{dataset}{Pareto front on which datatset}
  \item{solver}{Pareto front for which SVM solver}
  \item{repl}{number of replication}
  \item{error}{first performance measure - error = 1 - accuary}
  \item{execTime}{second performance measure - the training time}
}}
\usage{
apprSVMParetoFronts
}
\description{
A dataset containing the Pareto Fronts of differen approximative SVM sovlers
with respect to the objectives accuary and training time. A priori a multi-
objective parameter tuning has been done for every solver, the resulting
Pareto fronts of 10 independent optimizations runs on 4 data sets are given
}
\keyword{datasets}

