% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareCommonParetoFronts.R
\name{compareCommonParetoFronts}
\alias{compareCommonParetoFronts}
\title{Measure for the comparison of a true and an estimated common pareto front.}
\usage{
compareCommonParetoFronts(landscape, portfolio)
}
\arguments{
\item{landscape}{\cr
Result of \code{generateParetoLandscape}.}

\item{portfolio}{\cr
Result of \code{MOSAP::selectPortfolio}.}
}
\value{
[\code{numeric}]
 Value between 0 and 1.
}
\description{
Measure for the comparison of a true and an estimated common pareto front.
}

