% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/countInversions.R
\name{countInversions}
\alias{countInversions}
\title{countInversions}
\usage{
countInversions(x)
}
\arguments{
\item{x}{[\code{numeric}]\cr
Numeric vector.}
}
\value{
[\code{integer(1)}].
  number of inversions
}
\description{
Counts the number of inversions in the numeric vector x. An inversion occurs if
\eqn{x[i] > x[j]} and \eqn{i < j}.
}

