% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mainTestProcedure.R
\name{mainTestProcedure}
\alias{mainTestProcedure}
\title{Main Test Procedure}
\usage{
mainTestProcedure(formula, data, alpha, indicator = "hv",
  perm.test = "mean.invs", kappa = 1e-06, normalize = TRUE,
  ref.point = c(1.1, 1.1), lambda = 100, n = 1e+05)
}
\arguments{
\item{formula}{[\code{formula}] \cr
Formula to specify the variables: algorithms ~ var1 + var2 | repl}

\item{data}{[\code{data.frame}] \cr
Data for the test. Must contain all variables of the formula, all other
variables are ignored.}

\item{alpa}{[\code{numeric(1)}] \cr
Alpha value for the tests.}
}
\description{
Main Test Procedure
}

