% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mse.R
\name{my_mse}
\alias{my_mse}
\title{Find the index of first/last \code{TRUE} value in a logical vector.}
\usage{
my_mse(predicted, true)
}
\arguments{
\item{predicted}{[\code{numeric}]\cr
Numeric vector vector.}

\item{true}{[\code{numeric}]\cr
Numeric vector vector.}
}
\value{
[\code{numeric(1)}].
  MSE
}
\description{
Find the index of first/last \code{TRUE} value in a logical vector.
}

