% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mosap_result.R
\name{plot.mosap_result}
\alias{plot.mosap_result}
\title{Plot-function for \code{frontTestResult}-objects.}
\usage{
\method{plot}{mosap_result}(x, make.pause = TRUE, ...)
}
\arguments{
\item{x}{[\code{mosap_result}]\cr
Result object from function \link{selectPortfolio}.}

\item{make.pause}{[\code{logical(1)}]\cr
Should the process be paused after each iteration?
Default is \code{TRUE}.}

\item{...}{[\code{any}] \cr
Not used.}
}
\value{
Nothing.
}
\description{
Plot-function for \code{frontTestResult}-objects.
}

