% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotValidationData.R
\name{plotValidationData}
\alias{plotValidationData}
\title{Plots validation data for the algorithm portfolio selection.}
\usage{
plotValidationData(val.data, repl = 1L, grey = FALSE, return.plot = FALSE,
  title = "", legend = TRUE)
}
\arguments{
\item{val.data}{\cr
Result of \code{generateValidationData}.}

\item{repl}{[\code{integer}] \cr
Which replication should be plotted? Default is the first replication.}

\item{grey}{[\code{logical}] \cr
Should the plot be in grey shades? Default is \code{FALSE}.}

\item{return.plot}{[\code{logical}] \cr
Should the plot be returned by the function (or just plotted)? Default
is \code{FALSE}.}

\item{title}{[\code{character}] \cr
Title for the plot. Default is no title.}

\item{legend}{[\code{logical}] \cr
Should a legend be plotted? Default is \code{TRUE}.}
}
\description{
Plots validation data for the algorithm portfolio selection.
}

