% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print_frontTestResult.R
\name{print.frontTestResult}
\alias{print.frontTestResult}
\title{print test result}
\usage{
\method{print}{frontTestResult}(x, ...)
}
\arguments{
\item{x}{[\code{frontTestResult}]\cr
Result object from function \link{selectPortfolio}.}

\item{...}{[\code{any}] \cr
  Not used.}
}
\description{
print test result
}

