% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print_mosap_result.R
\name{print.mosap_result}
\alias{print.mosap_result}
\title{print test result}
\usage{
\method{print}{mosap_result}(x, ...)
}
\arguments{
\item{x}{[\code{mosap_result}]\cr
Result object from function \link{selectPortfolio}.}

\item{...}{[\code{any}] \cr
  Not used.}
}
\description{
print test result
}

