% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sortedVectorTest.R
\name{sortedVectorTest}
\alias{sortedVectorTest}
\title{Sorted Vector Test}
\usage{
sortedVectorTest(x, order = sort(unique(x)), n = 1e+05)
}
\arguments{
\item{x}{[\code{numeric | factor}]\cr
Vector.}

\item{order}{[\code{numeric | factor}]\cr
Vector specifying the order for the element in x. Must have the same type
as x. Default is for numeric vectors the numerical order, for factors
the order of the levels, for character vectors the first occurance in x.}

\item{[\code{integer(1)}]\cr}{Count of permutations used for the test.}
}
\description{
Is the input vector sorted with respect to the in order given order?
Verified via a permutation test:
H_0: The elements in the vector are randomly permutated
H_1: The element in the vector follow the order given in order.
}
\details{
n random permutations of x are drawn and for each permutation the number
  of inversions with respect to the given order is calculated. The rate of
  random permutations having less inversions than x is returned
}

