% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateValidationData.R
\name{generateValidationData}
\alias{generateValidationData}
\title{Generates validation data for the algorithm portfolio selection.}
\usage{
generateValidationData(N, M, split.points.type, type, randomness)
}
\arguments{
\item{N}{[\code{integer}] \cr
Number of algorithms on the common pareto front.}

\item{M}{[\code{integer}] \cr
Number of additional algorithms that are not on the commo pareto front.}

\item{split.points.type}{[\code{character}] \cr
Determines whether the split.points between the algorithms are chosen in  a 
\code{uniform} or \code{nonuniform} way.}

\item{type}{[\code{character}] \cr
Determines how the discrete approximation of the pareto front is done.
The values \code{deterministic}, \code{uniform}, \code{NSGA2} and \code{NSGA2_g}
are possible.}

\item{randomness}{[\code{character}] \cr
Determines whether noise is added to the parameters of the functions that form
the fronts (\code{parambased}) or to the points (\code{pointbased}) to get 
replications of the discrete approximation.}
}
\value{
[\code{list}]
 List that contains information about the true (original) pareto landscape 
 (functions that form the pareto fronts, true splitpoints, names of the algorithms)
 and the validation data.
}
\description{
Generates validation data for the algorithm portfolio selection.
}

