% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_frontTestResult.R
\name{plot.frontTestResult}
\alias{plot.frontTestResult}
\title{Plot-function for \code{frontTestResult}-objects.}
\usage{
\method{plot}{frontTestResult}(x, make.pause = TRUE, ...)
}
\arguments{
\item{x}{[\code{frontTestResult}]\cr
Result object from function \link{selectPortfolio}.}

\item{make.pause}{[\code{logical(1)}]\cr
Should the process be paused after each iteration?
Default is \code{TRUE}.}

\item{...}{[\code{any}] \cr
Not used.}
}
\value{
Nothing.
}
\description{
Plot-function for \code{frontTestResult}-objects.
}

