% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_frontTestResult.R
\name{renderFrontTestResult}
\alias{renderFrontTestResult}
\title{render test result plots (without eaf plots)}
\usage{
renderFrontTestResult(x, colors = NULL)
}
\description{
render test result plots (without eaf plots)
}

